refactor: 장바구니 목록 페이지 스타일, 타입 파일 분리
