+++ src/pages/ProductListPage/ProductList/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const ProductListContainer = styled.ul`
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  column-gap: 60px;
+  row-gap: 80px;
+  margin-bottom: 160px;
+`;
