+++ src/pages/ProductListPage/style.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  width: 60%;
+  min-width: 1024px;
+  margin: auto;
+`;
