+++ src/types/index.ts
@@ -1,0 +1,12 @@
+export type Product = {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+};
+
+export type ShoppingBasketProduct = {
+  id: number;
+  quantity: number;
+  product: Product;
+};
